Suggested changes made
